Multiply sprite dimensions by 0.7071 to...
...compensate for the maximum offset caused by the rotation of a sprite.
This causes the boids to wrap just as the square containing their sprite
rotated at 45 degrees goes off-screen.

If a square is not rotated at all, it won't wrap until it is at a point
slightly further than right over the edge of the screen; but this isn't
as noticeable as a rotated square wrapping too early.