Fix alignment
Previously the neighbor angles were combined to form an average angle,
but this causes the resultant angle to skew towards 180 degrees.

Now the combination of neighbor angles is done by converting the angles
into unit vectors, then combining those unit vectors and normalizing the
resultant vector, and finally taking the angle of the resultant vector.